move DefaultTableSource to datasource module
