+++ datafusion/core/src/physical_plan/planner.rs
@@ -25,9 +25,10 @@ use super::{
 use crate::execution::context::{ExecutionProps, SessionState};
 use crate::logical_expr::utils::generate_sort_key;
 use crate::logical_plan::plan::{
-    source_as_provider, Aggregate, EmptyRelation, Filter, Join, Projection, Sort,
+    Aggregate, EmptyRelation, Filter, Join, Projection, Sort,
     SubqueryAlias, TableScan, Window,
 };
+use crate::datasource::source_as_provider;
 use crate::logical_plan::{
     unalias, unnormalize_cols, CrossJoin, DFSchema, Expr, LogicalPlan,
     Partitioning as LogicalPartitioning, PlanType, Repartition, ToStringifiedPlan, Union,
